Update index.Rmd
Sadie this is AMAZING! I loved looking at it and totally want to do it with my favorite musicians.  I only had a few notes, which I added to the text with #Sadie- and then the note.  If you have questions about what I added, let me know!  Great work.